* Imported the db before creating the blueprint * fixed load_logged_in_user() to be psycopg syntax
